inverse kin always gives [0,360]

+++ include/inverse_kin.h
@@ -1,23 +1,24 @@
 #include <Arduino.h>
 #include <stdio.h>
+#include <cmath>
 
 typedef struct
 {
-    float x;
-    float y;
+    double x;
+    double y;
 } Coordinate;
 
 typedef struct
 {
-    float a1;
-    float a2;
+    double a1;
+    double a2;
 } MotorAngles;
 
 typedef struct
 {
-    float s1;
-    float s2;
+    double s1;
+    double s2;
 } MotorSteps;
 
 MotorAngles coord_to_angles(int, int);
-int positiveMod(int, int);
+double positiveMod(double, double);
